RESI fixed - since last Friday, file command produced empty files. Picta (portable) now produces large pictures with fonts treated to their best. installplugin now works with zip files from http or local system

+++ glib/glrender.h
@@ -285,7 +285,7 @@ public:
   /* the functions set current view for drawinf large picture x, y- curent quadraterial,
    res - required resolution; to be used in conjunction with GetPixels
   */
-  void LookAt(int x, int y, short res);
+  void LookAt(double x, double y, short res);
 
   static TGraphicsStyles* GetStyles();
